Create edit info screen base skeleton and landing page edit button functionality
