+++ lib/screens/patient_info_screen.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
-
 import '../widgets/nav_bar.dart';
+import '../widgets/dashed_separator.dart';
+import './edit_info_screen.dart';
 
 class PatientInfoScreen extends StatefulWidget {
   @override
@@ -146,7 +147,12 @@ class _PatientInfoScreenState extends State<PatientInfoScreen> {
                   height: 30.0,
                   width: 800.0,
                   child: ElevatedButton(
-                    onPressed: () {},
+                    onPressed: () {
+                      Navigator.pushReplacement(
+                        context,
+                        MaterialPageRoute(builder: (context) => EditInfoScreen()),
+                      );
+              },
                     child: Text('Edit'),
                   ),
                 ),
@@ -164,35 +170,3 @@ class _PatientInfoScreenState extends State<PatientInfoScreen> {
     ); //SafeArea
   }//build
 }//_PatientInfoScreenState
-
-class MySeparator extends StatelessWidget {
-  const MySeparator({Key? key, this.height = 1, this.color = Colors.black})
-      : super(key: key);
-  final double height;
-  final Color color;
-
-  @override
-  Widget build(BuildContext context) {
-    return LayoutBuilder(
-      builder: (BuildContext context, BoxConstraints constraints) {
-        final boxWidth = constraints.constrainWidth();
-        const dashWidth = 2.0;
-        final dashHeight = height;
-        final dashCount = (boxWidth / (2 * dashWidth)).floor();
-        return Flex(
-          children: List.generate(dashCount, (_) {
-            return SizedBox(
-              width: dashWidth,
-              height: dashHeight,
-              child: DecoratedBox(
-                decoration: BoxDecoration(color: color),
-              ),
-            );
-          }),
-          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-          direction: Axis.horizontal,
-        );
-      },
-    );//LayoutBuilder
-  }//build
-}//MySeparator
